Split math submod
